更新说明: 1.添加注释 2.优化查询对象,新增AbstractQuery 抽象类作为基类,减少重复编码
