+++ src/main/java/io/mycat/dao/DomainER/O2MQuery.java
@@ -1,16 +1,15 @@
 package io.mycat.dao.DomainER;
 
-import io.mycat.dao.DomainER.abs.AbstractQuery;
-import io.mycat.dao.query.AutoQueryConditonHandler;
+import io.mycat.dao.DomainER.abstractbean.AbstractQuery;
 import io.mycat.dao.query.DynaQueryCondHanlder;
-import io.mycat.dao.query.PagedQuery;
 import io.mycat.dao.util.NameUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Field;
-import java.util.*;
-import java.util.function.Function;
+import java.util.Collection;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 /**
  * 一对多查询对象
@@ -24,20 +23,16 @@ import java.util.function.Function;
  *
  * @author jim
  */
-public class O2MQuery  extends PagedQuery   {
+public class O2MQuery extends AbstractQuery {
     protected static Logger log = LoggerFactory.getLogger(O2MQuery.class);
-    private final List<QueryField> queryFields = new ArrayList<>();
-    private boolean autoRemoveDuplicateFields = true;
-    private DynaQueryCondHanlder condHandler = null;
-    private AutoQueryConditonHandler defaultConHandler = new AutoQueryConditonHandler();
-    private String groupBy = "";
+
     //存储子查询对象
     public Map<String, ChildrenDomainQuery> childrenDomainMap = new LinkedHashMap<>();
     //当前操作的子查询名称
     private String childrenDomanName;
 
     public O2MQuery() {
-        defaultConHandler.setTranslateFun(replaceQueryFieldByAlias());
+        defaultConHandler.setTranslateFun(super.replaceQueryFieldByAlias());
     }
 
     private O2MQuery addQueryFields(Collection<? extends QueryField> queryFields) {
@@ -90,100 +85,7 @@ public class O2MQuery  extends PagedQuery   {
     }
 
     public String buildSQLNoPage() {
-        Set<String> allDomains = new LinkedHashSet<>();
-        StringBuilder sb = new StringBuilder().append("SELECT ");
-        for (QueryField field : queryFields) {
-            if (field instanceof DomainField) {
-                DomainField theField = (DomainField) field;
-                allDomains.add(theField.domain.domainCls.getCanonicalName());
-
-            }
-            sb.append(field.getAlias()).append(",");
-        }
-        sb = sb.deleteCharAt(sb.lastIndexOf(",")).append(" from ");
-        for (String domainName : allDomains) {
-            DomainInfo domain = DomainAutoRelations.findDomainByClassName(domainName);
-            sb.append(domain.tableName).append(",");
-
-        }
-        sb = sb.deleteCharAt(sb.lastIndexOf(","));
-        sb.append(" WHERE ");
-        boolean hasJoin = allDomains.size() > 1;
-        if (hasJoin) {// 有JOIN
-            int joinCount = 0;
-            sb.append('(');
-            for (String domainName : allDomains) {
-                DomainInfo domain = DomainAutoRelations.findDomainByClassName(domainName);
-                for (String parentDomain : allDomains) {
-                    if (domain.parentTableAndForeinKey.containsKey(parentDomain)) {
-                        String foreinKey = domain.parentTableAndForeinKey.get(parentDomain);
-                        DomainInfo parent = DomainAutoRelations.findDomainByClassName(parentDomain);
-                        sb.append(domain.tableName + "." + domain.getField(foreinKey).dbColumn + " = "
-                                + parent.tableName + "." + parent.idColumn).append(" AND ");
-                        joinCount++;
-                    }
-                }
-
-            }
-            if (joinCount + 1 < allDomains.size()) {
-                throw new RuntimeException("Join Count is  " + joinCount + ",but Selected domains count "
-                        + allDomains.size() + " domains :" + Arrays.toString(allDomains.toArray()));
-            }
-            sb.delete(sb.lastIndexOf(" AND "), sb.length());
-            sb.append(')');
-
-        }
-        if (this.condHandler != null) {
-            String cond = condHandler.genCondtions(queryParams);
-            if (cond != null && !cond.isEmpty()) {
-                if (hasJoin) {
-                    sb.append(" AND ( 1=1 ").append(cond).append(" ) ");
-                } else {
-                    sb.append(cond);
-                }
-            }
-
-        } else if (!hasJoin) {
-            // 没有JOIN ，没有Where
-            sb.delete(sb.lastIndexOf(" WHERE "), sb.length());
-        }
-        if (!"".equals(this.groupBy)) {
-            sb.append(" " + this.groupBy);
-        }
-        if (this.orderBy != null) {
-            sb.append(" " + this.orderBy);
-        }
-
-        return sb.toString();
-    }
-
-    private Function<String, String> replaceQueryFieldByAlias() {
-        return (t) -> {
-            for (QueryField field : O2MQuery.this.queryFields) {
-                String alias = field.getAlias();
-                int index = t.indexOf(alias);
-                boolean found = false;
-                if (index == 0 || (index > 0 && t.charAt(index - 1) != ':')) {
-                    found = true;
-
-                } else if (index > 0) {
-                    index = t.lastIndexOf(alias);
-                    if (index == 0 || (index > 0 && t.charAt(index - 1) != ':')) {
-                        found = true;
-                    }
-                }
-                if (found) {
-                    return t.substring(0, index) + field.getSQLExpress() + t.substring(index + alias.length());
-                }
-
-            }
-            return t;
-        };
-
-    }
-
-    public List<QueryField> getQueryFields() {
-        return queryFields;
+        return super.buildSQLNoPage();
     }
 
     /**
