+++ src/main/java/io/mycat/dao/DomainER/QueryField.java
@@ -1,0 +1,11 @@
+package io.mycat.dao.DomainER;
+
+/**
+ * @author Leader us
+ */
+public interface QueryField {
+
+    String getAlias();
+
+    String getSQLExpress();
+}
