+++ src/main/java/io/mycat/dao/query/PagedQuery.java
@@ -18,6 +18,7 @@ import org.slf4j.LoggerFactory;
 public abstract class PagedQuery {
     protected static Logger log = LoggerFactory.getLogger(PagedQuery.class);
     protected String orderBy;
+    public String groupBy;
     protected Map<String, Object> queryParams = new HashMap<>();
     protected int pageIndex = -1;
     protected int pageSize = 20;
